adds authenticate decorator, adds validate uuid, updates users to add checks using these functions

+++ controllers/users_controller.py
@@ -2,10 +2,13 @@ from flask import jsonify
 from flask_bcrypt import generate_password_hash
 
 from db import db
+from lib.authenticate import authenticate
 from util.reflection import populate_object
+from util.validate_uuid4 import validate_uuid4
 from models.users import Users, user_schema, users_schema
 
 
+@authenticate
 def user_add(req):
     post_data = req.form if req.form else req.json
 
@@ -15,12 +18,22 @@ def user_add(req):
     new_user.password = generate_password_hash(new_user.password).decode("utf8")
 
     db.session.add(new_user)
-    db.session.commit()
+
+    try:
+        db.session.commit()
+
+    except:
+        db.session.rollback()
+        return jsonify({"message": "unable to add user"}), 400
 
     return jsonify({"message": "user added", "user": user_schema.dump(new_user)}), 201
 
 
+@authenticate
 def user_get_by_id(req, user_id):
+    if not validate_uuid4(user_id):
+        return jsonify({"message": "invalid user id"}), 400
+
     user_query = db.session.query(Users).filter(Users.user_id == user_id).first()
 
     if not user_query:
@@ -30,13 +43,18 @@ def user_get_by_id(req, user_id):
         return jsonify({"message": "user found", "user": user_schema.dump(user_query)}), 200
 
 
+@authenticate
 def users_get_all(req):
     users_query = db.session.query(Users).all()
 
     return jsonify({"message": "users found", "users": users_schema.dump(users_query)}), 200
 
 
+@authenticate
 def user_update_by_id(req, user_id):
+    if not validate_uuid4(user_id):
+        return jsonify({"message": "invalid user id"}), 400
+
     post_data = req.form if req.form else req.json
 
     user_query = db.session.query(Users).filter(Users.user_id == user_id).first()
@@ -49,12 +67,21 @@ def user_update_by_id(req, user_id):
     if post_data.get("password"):
         user_query.password = generate_password_hash(user_query.password).decode("utf8")
 
-    db.session.commit()
+    try:
+        db.session.commit()
+
+    except:
+        db.session.rollback()
+        return jsonify({"message": "unable to update user"}), 400
 
     return jsonify({"message": "user updated", "user": user_schema.dump(user_query)}), 200
 
 
+@authenticate
 def user_activity(req, user_id):
+    if not validate_uuid4(user_id):
+        return jsonify({"message": "invalid user id"}), 400
+
     user_query = db.session.query(Users).filter(Users.user_id == user_id).first()
 
     if not user_query:
@@ -70,13 +97,23 @@ def user_activity(req, user_id):
         return jsonify({"message": "user deactivated", "user": user_schema.dump(user_query)}), 200
 
 
+@authenticate
 def user_delete_by_id(req, user_id):
+    if not validate_uuid4(user_id):
+        return jsonify({"message": "invalid user id"}), 400
+
     user_query = db.session.query(Users).filter(Users.user_id == user_id).first()
 
     if not user_query:
         return jsonify({"message": "user not found"}), 404
 
     db.session.delete(user_query)
-    db.session.commit()
+
+    try:
+        db.session.commit()
+
+    except:
+        db.session.rollback()
+        return jsonify({"message": "unable to delete user"}), 400
 
     return jsonify({"message": "user deleted"}), 200
